# Conflicts:
#	Backend/cli-backend/src/main/java/org/example/models/Comment.java
#	Backend/cli-backend/src/main/java/org/example/models/Post.java
#	Backend/cli-backend/src/main/java/org/example/services/CommentService.java
#	out/production/Matcha/org/example/Main.class
#	out/production/Matcha/org/example/dbconnection/DatabaseConnection.class
#	out/production/Matcha/org/example/entities/ActionState$1.class
#	out/production/Matcha/org/example/entities/ActionState.class
#	out/production/Matcha/org/example/entities/Comment.class
#	out/production/Matcha/org/example/entities/Post.class
#	out/production/Matcha/org/example/repositories/CommentRepo.class
#	out/production/Matcha/org/example/repositories/PostRepo.class
#	out/production/Matcha/org/example/repositories/UserRepo.class
#	out/production/Matcha/org/example/services/CommentService.class
#	out/production/Matcha/org/example/services/PostService.class
#	out/production/Matcha/org/example/services/UserService.class
#	out/production/Matcha/org/example/userinterface/UIComment.class
#	out/production/Matcha/org/example/userinterface/UIPost.class

+++ Backend/cli-backend/src/main/java/org/example/repositories/UserRepo.java
@@ -21,6 +21,10 @@ public class UserRepo {
     }
 
     public void save(User user) throws SQLException {
+        if (!DatabaseConnection.isConnected()) {
+            return;
+        }
+
         String sql = "INSERT INTO profile (username, email, password) VALUES (?, ?, ?)";
         try (Connection conn = DatabaseConnection.getConnection();
              PreparedStatement pstmt = conn.prepareStatement(sql)) {
@@ -33,6 +37,9 @@ public class UserRepo {
     }
 
     public void load(ArrayList<User> users) throws SQLException {
+        if (!DatabaseConnection.isConnected()) {
+            return;
+        }
         String sql = "SELECT username, email, password FROM profile";
 
         Connection conn = DatabaseConnection.getConnection();
@@ -66,6 +73,10 @@ public class UserRepo {
     }
 
     public void deleteUser(String username) {
+        if (!DatabaseConnection.isConnected()) {
+            return;
+        }
+
         String sql = "DELETE FROM profile WHERE username = ?";
         try (Connection conn = DatabaseConnection.getConnection();
              PreparedStatement pstmt = conn.prepareStatement(sql)) {
